v1.1.0
-Added Market.
The Market is A server-wide Market that allows Players to buy and sell Items.
The Market's trades are edited by config, allowing server admins to create trades that every Player can see.
There is an example config file showing how it works here:

-Added Coin Stack.
There is a coin stack for every type of Coin.
They can be placed by holding 8 or more coins.
They show up in caves and are pretty uncommon.
Their generation rarity value and the ability to disable their generation are in the config.

-Added Cheap Money Bag * Rich Money Bag.
The Money Bags contain a random amount of Coins when opened.
They are currently not obtainable through crafting/loot.
Instead, this Item could be placed in the Market or as a quest reward from another mod.
Their random range values are in the config.

--Added the ability to broadcast a Trading Post.
There is a button in the Trading Post GUI that, when pressed, sends a message to everyone on the server containing information about the trade.
There is a delay when the button is pressed (currently 10 seconds) that the player must wait untill broadcasting again, preventing major spam.
The broadcast delay value and the ability to disable broadcasts are in the config.

--Added Coin sound effects for collecting Coins, trading in the Market, and opening Money Bags.

-Changed the location of the config file!
If you want to keep your old config values, please make sure you transfer your old file into the folder located at "config/CalemiUtils/".

-Changed the "/cu" command to "/cutils".
The "/cu" command was conflicting with some mods.

-Fixed Trading Post not working well with certain stack sizes.
-Fixed Trading Post not working well with certain NBTs.
-Fixed Trading Post's overlay not showing some info.

-Refined code by more organizing and comments.